tasks: Install ruff
In our projects we have used ruff long enough to commit to it. Let's
have it in our development toolboxes by default then, so that we don't
keep sending PRs which break ruff rules.

This also enables our other projects to start using ruff in CI (which
runs in tasks, instead of a dedicated unit-tests container).