tasks: Disable installing recommends
firefox' Recommends pull in mesa and half of a desktop stack (like
tracker, upower, and mesa), and thus adds an unnecessary 600 MB of
dependencies.

Other packages pull in nodejs-full-i18n, various -docs, systemd-networkd
and other bits which have no place in a container.

Explicitly install passwd to keep the container working with toolbox,
and adobe-source-code-pro-fonts to keep a proper monospace font (for
pixel tests).