tasks: Reset SELinux level of images on startup
In some situations (in particular the fedora-rawhide-boot and -payload
images) gain an extra SELinux "Multi-category range" context like this:

    system_u:object_r:container_file_t:s0:c220,c230  fedora-rawhide-boot-ba19e37f2b38f6578574275b7eb209e5b1850bceb9eb671cb366db16e4bf78fc.iso

This makes them undeletable by `image-prune`, `rm`, or anything else
that the container could do. This causes old images to pile up, and the
host eventually fails with ENOSPC.

There is nothing that the container can do to fix these files, and it's
not clear what causes this in the first place. To mitigate, reset the
context level in the system unit at startup.