tasks: Bind the host's podman API socket
This paves the way for spawning per-job tasks containers from the
container (via `job-runner`).

Getting the permissions right is unfortunately annoyingly complicated, as the
host's socket has 660 permissions, but the tasks container runs as uid 1111.
Ideally we could use something like

    -v "${XDG_RUNTIME_DIR:-/run}/podman/podman.sock:/podman.sock:idmap=gids=$(id -g)-1111-1"

but that fails with "invalid mappings", and is generally not well documented.
`--mount=type=bind,[...],idmap --uidmap [...]` does not work either.

So resort to file permissions for the host's `podman.sock`:
 - For local developers in `run-local.sh`, make the socket world
   writable. This does not actually hurt for a human developer: Its
   directory (/run/user/uid) is not accessible for any other user.

 - For production, set the socket group to `1111`. That doesn't matter
   much, the secrets are all already chmod'ed to the container user, and
   these machines don't do anything else. `setfacl` would be more
   targeted, but it isn't installed in Fedora IoT/CoreOS.