tasks: Drop cockpit-tasks as main entry point
We already overwrite the entry point in the webhook container and on
OpenShift, and we are going to for the impending job-runner/job split.

So drop it from the image, and instead set it explicitly in the systemd
unit.

Rework run-local.sh accordingly: Replace the async "wait for
initialization" loop with a synchronous bots clone, and only run the
`cockpit-tasks` main loop for `test_pr()` (so that we keep covering that
script) -- the other tests don't need it. This paves the way for adding
a test which covers `run-queue` with custom parameters (against a mock
GitHub API), which we will introduce in the next commit.