tasks: Move npm configuration to static file
Create a static /etc/npmrc on our tasks bot hosts which is suitable for
machines in the internal Red Hat network. (The role already assumed that
before, so that's not a regression - and easy enough to conditionalize
if we need to in the future).

This gets rid of the `npm config` calls in `setup-tasks`.