tasks: Eliminate setup-tasks
The bots GitHub API recently added `$COCKPIT_GITHUB_TOKEN_FILE` [1], which
allows us to not create the ~/.config/cockpit-dev/github-token symlink any
more. It already supports `$COCKPIT_S3_KEY_DIR`. So set these variables in
the tasks container.

Adjust the OpenShift secrets build to not consider subdirectories. This
was only relevant for the s3-keys/ subdirectory -- but we don't
currently need them on our CentOS CI deployment. If we ever do, we'll
create a separate volume out of that directory. This gets rid of the
`--` subdirectory encoding.

These were the last things in `setup-tasks`, so remove it.

[1] https://github.com/cockpit-project/bots/pull/6016