tasks: Drop unnecessary resources from cockpit-tasks@* containers
These now just do `{tests,issue}-scan`, `run-queue`, `job-runner`, and
test statistics. None of these uses a lot of resources, the expensive
ones (tests and image refreshes) moved to job-runner.toml. So drop all
the memory/shm configuration and obsolete env vars.

Also drop the custom network. If inter-container socket leakage is still an
issue, it'll need to be done to the job containers, but hopefully that was
fixed long ago.🤞

Drop the remaining `Environment=` indirection. The env vars can just be
expanded in-place when writing the service file. This was already done for many
of them.

Still keep importing /dev/kvm into the container. It doesn't strictly
need this, but current bots `run-queue` uses the presence of this as an
indication whether or not it can consume tasks from the public/rhel
queues. This needs to be fixed first (which needs some thought). So
dropping that will happen later.