tasks: Test posted statuses in mock PR and image refresh
Teach `mock-github` to write a log of POST requests, and add some
initial assertions. They are not very precise yet (that'll get cleaned
up in the Python rewrite soon), but at least ensure that we post the
statuses to the expected SHA.

This enables us to create an integration test for a cross-project test
in a future commit.