tasks: add python3-mypy, pytest modules
We're going to stop using toxbox to run our static checks against a venv
and start depending on the run from the tasks container instead.  We
need mypy in the tasks container for that, of course.

While we're at it, add python3-pytest.  Something else we're already
using (probably aioresponses) is pulling that in implicitly, but it's
pretty important for us, so make sure we have it explicitly.

Add a couple of more useful pytest modules as well...