ansible: Add SELinux policy for podman socket
When the tasks/job containers access the bind-mounted podman.sock, they
currently run into a lot of SELinux denials. This happened to work
as we have run SELinux in permissive mode on our bots in the last two
weeks (see commit c118069e5476).

Create a proper SELinux policy plugin instead, so that we can put it
back into enforce mode.