workflows: build a multi-arch tasks container
Build a tasks container for arm64 in addition to amd64.

We do this by splitting the build process into multiple jobs.  For the
arm one, we make use of the arm runners from BuildJet.  At the end, we
write the manifest and upload that with the "public" tag name.

Remove the local infrastructure and rules for pushing the container;
it's not so easy anymore.