ansible: Fix sync-secrets playbook to not clobber s3-keys alias symlinks
The psi/image-cache playbook installs the s3-keys/ symlink for the local
S3 mirror on rhos-01-1. But sync-secrets clobbers that (as it always
removes the entire existing secrets dir before re-syncing it).

Fix this by factoring out the symlink creation into a nwe
"local-s3-alias" role and calling it from both places.