ansible: Configure persistent git cache volume for job containers
Define a `cockpituous-git-cache` persistent volume so that the expensive
and big node-cache.git (plus the much less expensive bots.git) get
re-used between jobs.

Note that we don't want to keep *all* of `$XDG_CACHE_HOME`, in
particular not the libvirt/ directory. So restrict that to the
`cockpit-dev/` subdirectory, which is handled by cockpit's
`git-utils.sh`.

Fixes #611