tasks: Delay webhook startup
Starting the AMQP and webhook container in parallel very often runs into
webhook not yet being able to connect to the AMPQ server, and thus
getting trapped in a crash loop until it wins the race.

Delay the webhook startup by 10 seconds. This isn't scientific, but good
enough for the two times a year that we restart this pod.