webhook: Retry OSError on queue connection failure
Commit 06c2717be74 introduced a retry loop for connecting to AMQP, as
these two containers start at the same time in the OpenShift deployment.

However, these days the webhook often fails with `ssl.SSLEOFError`
(with parents `SSLError` and `OSError`), which doesn't fit `AMQPError`.
So intercept this as well.

With that we can revert the static sleep from commit 169b53bac1998
again.

Fixes #635