cockpit-tasks: poll more often
When we used to hammer GitHub to pick up tasks to run, it made more
sense to be "gentle" but now that we're just polling our own AMQP queue,
we can be a lot less cautious.

Instead of sleeping a random interval between 1 and 10 minutes, just
sleep 10 seconds.  This will result in substantially reduced latency
when picking up tasks and will also result in better distribution of
very short tasks across machines (as may occur when a task fails).