local-s3: Switch to Hummingbird minio containers
The official image is deprecated and not updated any more [1][2], so we need to
switch to a maintained image. Project Hummingbird just got a new one [3], so
switch to that.

This requires some adjustments to volume paths, as the Hummingbird image runs
as non-root user and thus does not have `HOME=/root`.

Related: https://issues.redhat.com/browse/HUM-167

[1] https://github.com/minio/minio/commit/9e49d5e7a648f00e
[2] https://github.com/minio/minio/issues/21647
[3] https://quay.io/repository/hummingbird/minio